Moved components into correct folder
