+++ client/src/components/details/header.tsx
