+++ client/src/components/details/linkbox.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 export function LinkBox({links}: { links: string[] }) {
     return (
         <div className="bg-white bg-opacity-10 py-8 px-32 rounded-md mb-8 mt-2">
